backend/requirements.txt
Django>=4.2,<5.0
gunicorn
psycopg2-binary
dj-database-url
djangorestframework
django-cors-headers
Pillow
whitenoise
djangorestframework-simplejwt
